Fix parsing error

+++ lib-client/src/main/java/com/baptistecarlier/kotlin/datagouvfr/client/models/CommunityResource.kt
@@ -49,7 +49,7 @@ data class CommunityResource(
     //@SerialName("extras")
     //var extras: Map<String, Any?>? = null,
     @SerialName("filesize")
-    var filesize: Int? = null,
+    var filesize: Long? = null,
     @SerialName("id")
     var id: String? = null,
     //@SerialName("last_modified")
